+++ src/main/java/com/mykolas/ignitismessagetask/statistics/StatisticsQueries.java
@@ -19,12 +19,12 @@ public class StatisticsQueries {
     @Autowired
     private DSLContext create;
 
-    public Integer fetchTotalNumberMessagesByAuthor(Integer authorId) {
+    public Integer fetchTotalNumberMessagesByAuthor(Long authorId) {
         return create
                 .fetchCount(Tables.MESSAGES, Messages.MESSAGES.AUTHOR_ID.eq(authorId));
     }
 
-    public Integer fetchTotalMessagesByReceiver(Integer receiverId){
+    public Integer fetchTotalMessagesByReceiver(Long receiverId){
         return create
                 .fetchCount(Tables.MESSAGES, Messages.MESSAGES.RECEIVER_ID.eq(receiverId));
     }
@@ -33,7 +33,7 @@ public class StatisticsQueries {
        return Objects.requireNonNull(create
                .select(DSL.min(Messages.MESSAGES.TIME))
                .from(Tables.MESSAGES)
-               .where(Messages.MESSAGES.AUTHOR_ID.eq(Math.toIntExact(authorId)))
+               .where(Messages.MESSAGES.AUTHOR_ID.eq(authorId))
                .fetchOne()).value1();
     }
 
@@ -41,12 +41,12 @@ public class StatisticsQueries {
         return Objects.requireNonNull(create
                 .select(DSL.max(Messages.MESSAGES.TIME))
                 .from(Tables.MESSAGES)
-                .where(Messages.MESSAGES.AUTHOR_ID.eq(Math.toIntExact(userId)))
+                .where(Messages.MESSAGES.AUTHOR_ID.eq(userId))
                 .fetchOne()).value1();
     }
 
 
-    public Double fetchAverageLengthOfMessageByAuthorId(Integer userId){
+    public Double fetchAverageLengthOfMessageByAuthorId(Long userId){
 
        BigDecimal avgValue =  create
                 .select(DSL.avg(Messages.MESSAGES.LENGTH))
@@ -65,7 +65,7 @@ public class StatisticsQueries {
        return Objects.requireNonNull(create
                .select(Messages.MESSAGES.CONTENT)
                .from(Tables.MESSAGES)
-               .where(Messages.MESSAGES.AUTHOR_ID.eq(Math.toIntExact(userId)).and(Messages.MESSAGES.TIME.eq(latestMessageTime)))
+               .where(Messages.MESSAGES.AUTHOR_ID.eq(userId).and(Messages.MESSAGES.TIME.eq(latestMessageTime)))
                .fetchOne()).value1();
 
     }
@@ -74,7 +74,7 @@ public class StatisticsQueries {
         return Objects.requireNonNull(create
                 .select(DSL.min(Messages.MESSAGES.TIME))
                 .from(Tables.MESSAGES)
-                .where(Messages.MESSAGES.RECEIVER_ID.eq(Math.toIntExact(receiverId)))
+                .where(Messages.MESSAGES.RECEIVER_ID.eq(receiverId))
                 .fetchOne()).value1();
     }
 
@@ -82,11 +82,11 @@ public class StatisticsQueries {
         return Objects.requireNonNull(create
                 .select(DSL.max(Messages.MESSAGES.TIME))
                 .from(Tables.MESSAGES)
-                .where(Messages.MESSAGES.RECEIVER_ID.eq(Math.toIntExact(receiverId)))
+                .where(Messages.MESSAGES.RECEIVER_ID.eq(receiverId))
                 .fetchOne()).value1();
     }
 
-    public Double fetchAverageLengthOfMessageByReceiverId(Integer receiverId){
+    public Double fetchAverageLengthOfMessageByReceiverId(Long receiverId){
 
         BigDecimal avgValue =  create
                 .select(DSL.avg(Messages.MESSAGES.LENGTH))
@@ -105,7 +105,7 @@ public class StatisticsQueries {
         return Objects.requireNonNull(create
                 .select(Messages.MESSAGES.CONTENT)
                 .from(Tables.MESSAGES)
-                .where(Messages.MESSAGES.RECEIVER_ID.eq(Math.toIntExact(receiverId)).and(Messages.MESSAGES.TIME.eq(latestMessageTime)))
+                .where(Messages.MESSAGES.RECEIVER_ID.eq(receiverId).and(Messages.MESSAGES.TIME.eq(latestMessageTime)))
                 .fetchOne()).value1();
     }
